requirements.txt
django
djangorestframework
#uwsgi
gunicorn
